Add files via upload
Project: Contact List In C++ 

Please scroll down and click on the download button to download the Contact List project for free

Contact List is based on the concept of recording contact details. This whole system is in the C++ language. Also, the user can perform all the tasks like adding, viewing, modifying, and deleting the contact records of students. There’s no login system for this project.

This is a simple student contact details management system which developed using C++. It utilizes file handling and shows the class and object of the programming language. This project will teach you how to add, list, modify, and delete records in the file in the C++ language. It is a very effective and simple understanding of beginners. This is a simple mini project in C++.

About System
The Contact List project is designed in ‘C++’ language. This system is based on a concept to store and generate all the records of the student’s Contact. Also, this program is considered as a simple database of contacts in a school where the user can add Contact details safely and it’s not time-consuming. This System makes it easy to store records of each. Moreover, the user can add personal details, view a list of it, modify and remove the details. Also, this mini project is easy to operate and understand by the users.

How To Run the Project?
In order to run the project, you must have installed Dev C++ or Code Blocks on your PC. (we recommend these IDE).